Move portrait match to integration
